requirements.txt
appdirs==1.4.2
certifi==2017.4.17
coreapi==2.3.0
coreschema==0.0.4
defusedxml==0.4.1
Django==1.8.18
django-allauth==0.30.0
django-appconf==1.0.2
django-braces==1.4.0
django-etc==0.9.2
django-extra-views==0.8.0
django-imagekit==4.0
django-jalali==2.4.2
django-jalali-date==0.1.9
django-jet==1.0.6
django-material==0.13.1
django-model-utils==3.0.0
django-suit==0.2.25
django-telegrambot==1.0.1
djangorestframework==3.5.3
djoser==0.5.4
future==0.16.0
google-api-python-client==1.6.2
googlemaps==2.4.6
httplib2==0.9.2
itypes==1.1.0
jdatetime==1.8.1
Jinja2==2.9.6
jsonfield==2.0.1
MarkupSafe==1.0
oauth2client==4.0.0
oauthlib==2.0.1
packaging==16.8
pilkit==2.0
Pillow==3.4.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-telegram-bot==6.0.1
python3-openid==3.0.10
pytz==2016.4
requests==2.11.1
requests-oauthlib==0.7.0
rsa==3.4.2
simplejson==3.6.5
six==1.10.0
sorl-thumbnail==12.3
uritemplate==3.0.0
